Updated Front End
Updated the front end for the application.

-Added Material UI
-Created Header Component

TODO:
-- Update front end to latest React Version
-- Create pagination for front end
-- Update components

This front end version served as a very helpful proof of concept for the overall functionality of the application. However it is not going to reflect the final MVP.
Over the next week or so I will build out a completely new front end for the application and work on creating the LEARN section. This will be a written tutorial on the various components of the blockchain, blockchain security, and how the blockchain operates.